requirements.txt
tinytuya>=1.13.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pyyaml>=6.0
python-dotenv>=1.0.0
httpx>=0.26.0
